services/product-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.1
pydantic==2.1.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
elasticsearch==8.8.0
redis==4.6.0
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1
